+++ BusinessEntity/MiniApps/UserMiniApp/Contracts/Dtos/UserPropertyTypeEnum.cs
@@ -6,5 +6,6 @@ public enum UserPropertyTypeEnum
     Undefined = 0,
     RichDocBookmarks = 100,
     RichDocDisplayedLevelProperty = 101,
-    DocPrintPresets = 200
+    DocPrintPresets = 200,
+    TreeExpansionState = 300
 }